requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.8.3
distro==1.9.0
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jiter==0.11.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lorem-text==3.0
openai==1.54.4
packaging==25.0
pluggy==1.6.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.0
pytest-django==4.11.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sniffio==1.3.1
socksio==1.0.0
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
